Add PollSockets function.

+++ aether/adapters/modems/bg95_at_modem.cpp
@@ -87,8 +87,8 @@ void Bg95AtModem::Start() {
 
 void Bg95AtModem::Stop() {}
 
-void Bg95AtModem::OpenNetwork(std::uint8_t const context_index,
-                              std::uint8_t const connect_index,
+void Bg95AtModem::OpenNetwork(std::uint32_t const context_index,
+                              std::uint32_t const connect_index,
                               ae::Protocol const protocol,
                               std::string const host,
                               std::uint16_t const port) {
@@ -99,19 +99,19 @@ void Bg95AtModem::OpenNetwork(std::uint8_t const context_index,
   AE_TELE_ERROR(kAdapterSerialNotOpen, "Port {}", port);
 }
 
-void Bg95AtModem::CloseNetwork(std::uint8_t const context_index,
-                               std::uint8_t const connect_index) {
+void Bg95AtModem::CloseNetwork(std::uint32_t const context_index,
+                               std::uint32_t const connect_index) {
   AE_TELE_ERROR(kAdapterSerialNotOpen, "Context index {}", context_index);
   AE_TELE_ERROR(kAdapterSerialNotOpen, "Connect index {}", connect_index);
 }
 
-void Bg95AtModem::WritePacket(std::uint8_t const connect_index,
+void Bg95AtModem::WritePacket(std::uint32_t const connect_index,
                               std::vector<uint8_t> const& data) {
   serial_->WriteData(data);
   AE_TELE_ERROR(kAdapterSerialNotOpen, "Connect index {}", connect_index);
 };
 
-void Bg95AtModem::ReadPacket(std::uint8_t const connect_index,
+void Bg95AtModem::ReadPacket(std::uint32_t const connect_index,
                              std::vector<std::uint8_t>& data,
                              std::int32_t timeout) {
   //std::size_t size{};
